Added a complex asset creation tests for multi typed entity links.
This test is in response to a bug fix that went in a couple of months
back, a real tricky bug, but unfortunately no test got added. This test
covers the case where you have an entity (for example a workspace
entity) which has a sg_entity single link field which can point to both
assets and shots. If you run folder creation on this, the algorithm
should be able to correctly distingiush between items linked to shots
and items linked to assets.