Don't include newlines when reading linked core entry
If the linked core file (core_Linux.cfg) is edited by hand a newline
is often introduced by the editor. The code that reads this file uses
readline() which preserves the end-of-line character(s) if present and
these then get looked up as part of the filename. Stripping them first
solves this.

Signed-off-by: Alan Dann <alan.dann@shotgunsoftware.com>