Added handling of folder creation race condition edge case in core hook.
On certain network file storages, it is possible to end up with a race
condition when multiple users or processes create folders at the same
time, resulting in a "file exists" error, despite the fact that the core
hook starts by checking that the file doesn't exist. This change
intelligently looks for this error and ignores it should it occur.

Thanks to Ben De Luca at Rushes for this patch (support ticket 2951).