#24865 Added associated_entity_type parameter to task and step
When creating a step or task folder, toolkit needs to figure out which
parent folder to associate the step or task with. This is currently done
by looking up the folder tree and picking the first shotgun folder that
is found. However, For complex folder setups, where for example you want
to have a structure with Sequence > Shot > Department > Step, this
approach won't work, because it will find Department rather than the
desired Shot level.

This change adds an optional associated_entity_type parameter to
the folder configs which allows a user to explicitly specify the type of
the parent folder to bind the task or step to. In the example above, we
would set associated_entity_type: Shot in order to tell the system not
to associate the step with the department object but instead (correctly)
with the Shot object.

Closes #61.