#23778 Improved QT Styling Adds a new method to the engine base class: _initialize_dark_look_and_feel() which replaces the _get_standard_qt_stylesheet() method (which will still be around however deprecated).
This method sets up a visual environment very similar to the one in
maya. It is using a slightly different approach than just using a
stylesheet - instead, it is a combination of style, palette and some css
on top. The hope is that this should provide a broad, solid and
consistent style base for all apps and widgets and ensure that they are
looking similar to the way they would look in Maya (or Nuke for that
matter).

Engine initialization will typically look something like this with the
new method:

# start up our QApp now
qt_application = QtGui.QApplication([])
qt_application.setWindowIcon(QtGui.QIcon(self.icon_256))
self._initialize_dark_look_and_feel()

Closes #80.