#25699 Improved backwards compatibility for frameworks.
When we added support for auto-updating frameworks (with 1.x.x style
version numbers), we omitted a backwards compatibility edge case; we
started to assume that the framework instance name in the environment
file was on the form frameworkname_v1.2.3 (or _v1.x.x for the new
syntax). This was not a problem for our own code since we had been
following this convention - it was however technically possible to
define a framework instance and name if foo - the old code would
correctly find it (based on its version and name) where the new code
expected a certain instance name. This turned out to be problematic for
one of our clients and this fix resolves that issue, effectively making
the system backwards compatible. 

Closes #85