#25757 Added tank commands to "unlocalize" core.
This adds two new commands which allows for a localized configuration to
be repointed to an external core:

share_core
----------
When new projects are created, these are often created in a state where
each project maintains its own independent copy of the core API. This
command allows you to take the core for such a project and move it out
into a separate location on disk. This makes it possible to create a
shared core, where several projects share a single copy of the Core API.
Note: if you already have a shared Core API that you would like this
configuration to use, instead use the attach_to_core command.

attach_to_core
--------------
When new projects are created, these are often created in a state where
each project maintains its own independent copy of the core API. This
command allows you to attach the configuration to an existing core API
installation rather than having it maintain its own embedded version of
the Core API. Note: If you don't have a shared core API yet, instead use
the share_core command.

closes #94