#25760 #25759 Project setup refactor, auto-path and storageless setups
This change contains several larger functionality tweaks:

- A new mode called 'auto_path' has been added to setup_project. This
mode is available via the API only and allows a project to be set up in
such a way that no pipeline configuration path entries are associated
with the project. This means that the project is not accessible from a
studio level tank command nor from within shotgun. The only way to
access it is via its built-in tank command. This is an intermediate mode
which we are putting in with the expectation of a coming change which
will allow entire configurations to be zipped up, uploaded and managed
from within shotgun itself, a so called cloud based config.

- Toolkit now supports the notion of a project without any storages.
Previously, a project needed to at least have a primary storage. A
storage less project does not have an associated project data location
and therefore no path cache. Some toolkit functionality is also not
relevant, for example folder creation, and will raise error messages if
you try to execute it on a project which doesn't have storages defined.

- The project setup structure has been extensively refactored and
reshuffled in order to create a new accessor in the form of a wizard
interface, which will be used by our upcoming shotgun desktop workflows.
The functionality of all existing commands and APIs remain the same
(apart from some tweaks in the output these commands generate).

Closes #90.