#26162 Minor changes to the tank setup_project localize logic.
This tweaks the logic for the tank setup_project command so that if the
currently running core API belongs to a localized pipeline
configuration, the created pipeline configuration will be localized too.
This brings it more in line with the way the desktop wizard project
setup process works and will be relevant going forward, as more and more
projects will be created as localized by default, as UI wizard adoption
will start becoming more commonplace.

Closes #98