#26551 Tank cmd: Project ctx now takes precendence for unknown locations
If you are using a project specific tank command, the project with which
you are being associated is implied by the location of the tank command.

If you do tank /tmp/foo/bar, a project specific tank command currently
resolves to the Empty Context, because the path /tmp/foo/bar does not
match any project path. With this change, the tank command will fall
back and default to a project context rather than an empty context in
the case a context cannot be determined.

Closes #104.