#27253 Fixes regression in with cloned pipeline configuration init.
In 0.15.x, the pipeline configuration data is cached in a new way,
eliminating the need for one of the lookup files in the project storage
location. There was, however, a regression caused by the fact that the
cache was not refreshed from Shotgun unless there was a full cache miss
- e.g. the entire project was missing from the pipeline configuration.
If there was a primary config present for a project in the cache, the
cache was deemed as being valid. This meant that if someone cloned the
primary config, the data about the clone would not appear in the cache
and toolkit would think that the clone location was not a valid pipeline
configuration.

This change fixes this issue by adding a higher-level try/catch, making
it more consistent with the behaviour in the sgtk.from_entity() method.

Closes #115.