#27476 Localize command now tries to switch more intelligently.
Previously, the tank localize command would first copy and set up the
core api for a pipeline configuration and once that was done it would
copy all engines, apps and frameworks across. This caused a potentially
extended period of time where the pipeline configuration was in a
half-completed state and wasn't operational.

This fix moves things around so that the apps and engines are copied
first, and then the API is copied. While not a perfect fix, this should
hopefully still resolve most issues and minimize the time when the
pipeline configuration exists in an in-between state.

Closes #119.