#27655 Fixed a bug causing path syncing to fail for retired storages.
When a Local File Storage is retired in Shotgun, attachment values
returned by the API that were previously referring to that storage get
truncated - the storage is no longer associated and paths are
consequently not constructured. The path cache sync logic previously
assumed that all file link data was complete and on the same form and
didn't take these truncated field values into account. This fix adds
support for this.