#26335 Tweaks to logic when running setup_project from a localized API
Previously, the logic would be trying to suggest a config location based
on the location of the core API. This worked well with our standard
setup where a 'studio' api was shared across projects. However, with the
desktop, things are more localized, and these locations may also be in 
%APPDATA% or Library locations, meaning that it doesn't make sense to
base new config default on them. This fix removes the default logic
whenever the core that is used to run the setup from is localized rather
than shared.

Closes #124