#27811 Tweaks to UI sidebar layouts to handle when parameters are long.
Following QA of #27811, this adds a horizontal scrollbar to make sure 
that long parameters without spaaces (e.g. template paths for example)
are displayed in a sensible fashion. Also makes some of the text
elements selectable which can be handy when you need to cut and paste
info out of the dialog.