#28334 Removed caching of environments to address reload problems.
It turns out that the environment class returns bundle setting
dictionaries by reference rather than by value (or as copies), meaning
that the environment object isn't really immutable, leading to potential
issues when caching it. We are seeing this with multi publish, where it
looks as if config path values are written back to the settings data in
the environment, causing the multi publish to fail to load when it is
being reloaded for the same environment again (e.g. a 'dev' reload).
While the optimization is still highly relevant, we are taking it out
for the time being so that it can be addressed fully at a later point.