Adding the code changes that i did as part of my support investigation.
This solves the issue with the json error outlined in the zendesk
ticket, however it only highlights that the current algorithm doesn't
allow for arbitrary sub-step levels to be defined. The Step processing
never triggers because in order to have the step processed, you need a
task to drive it.