#30136 App stylesheet handling by core
This adds logic to the engine so that every time a dialog is opened,
core will look for a style.qss file in the app's root directory, and if
present, apply this to the top level dialog widget. This makes it really
easy to drop in a qss file into any app. And if you have made changes
and want to reload, just re-open the dialog. No need to restart the
engine.

Furthermore, it introduces some standard colors that can be be
referenced in stylsheets:

- 'SG_HIGHLIGHT_COLOR': '#18A7E3'
- 'SG_FOREGROUND_COLOR': '#C8C8C8'
- 'SG_ALERT_COLOR': '#FC6246'

In order to reference them in a style.qss file, use a double bracket
syntax, e.g. {{SG_ALERT_COLOR}} and it will be replaced at runtime with
the right values. These values are also accessible via the
bundle.style_constants property.

Closes #187