#29962 Support multiple values passed in for deferred creation Adds support for passing multiple defer keywords to folder creation as a comma-separated string. This is useful if you want to trigger the deferred folder creation for multiple keywords when using the Launcher app to launch a DCC.
For example, when you launch Nuke, the Launcher app automatically sends the engine name (`tk-nuke`) as the deferred keyword, so any folders in your schema definition that are configured with the setting `defer_creation: "tk-nuke"` will be created. However, if you wanted this to also trigger folder creation for additional keywords, previously this was not possible because when you specify a specific keyword in the Launcher settings (eg. `tk-myApp`), this will override the app's default behavior of sending the engine name and replace it with the value you provide. Now you can send a defer keyword like `"tk-nuke, tk-myApp"`. Toolkit will treat that as 2 separate keywords (`tk-nuke` and `tk-myApp`) and will compare each one to the defer_creation setting for the folder.