#31543 sanitize incoming text input to UTF-8 in the setup project wizard
Since this code can be wrapped by external tools and scripts, we don't know where the input is coming from and it could be coming in as a `str` or `unicode` object. Toolkit generally assumes that all text is a UTF-8 encoded `str` object. Since this is somewhat of a bridge between Toolkit and external input (which could be from a custom script or tool that wraps this code), we ensure the input that comes in gets encoded correctly for Toolkit to use from here on in.