Remove panel matching logic and rename function
Panels should register commands, so we do not need to add a call to
match panels - we should use the commands registration instead.

The commands matching function is now called "get_matching_commands" and
the logic to fetch the setting is expected to be done before calling the
function. This makes is so that this function has only one purpose (or
at least a smaller one).