Fix some project actions not being properly cached
This is because the process_manager.py code from the websocket server
that calls this command extracts the entity_type from the env file (e.g.
shotgun_task.yml), which is all lowercase. Thus, shotgun_project.yml
would pass an entity_type of "project", which wouldn't match with
"Project". The result is that the "Check for core/app upgrades" commands
would not be cached, and thus would not be showed.