Start the shotgun engine with a specific context
When the entity_type is one of a task, we create a context for a
specific context (hardcoded for now - will be passed as args later).
This allows the shotgun_engine to have a context to pass to its apps
(e.g. applaunchers).

Having a specific context allows the applaunchers to properly expand
their icon paths which rely on a {target_engine} token. This token
fetches the engine path. The process of doing so eventually calls the
pick_environment hook. This hook uses the context to determine the
environment to choose. By passing a proper context (rather than an empty
context), we are able to resolve an environment that will be used to
fetch our target_engine path.