Cache the icon path as well
Also reduce the amount of code to fetch the proper parameters from the
command params.

This does not break the website (which also reads that cache), because
the website does not assume a fixed amount of tokens per line in the
cache, it just accesses the tokens that it uses (e.g. token[1]).