Implement get_entity_commands tank command
Calls the tank command of another pipeline config to get the commands
out of its cache (a new, more detailed cache), or update it if needed.

We can not reuse the Shotgun caches, as the Shotgun website interacts
with them with a constraining interface (it does not specify *what*
entities it wants the commands for). Not having an entity ID leaves us
with an empty context, which in turn makes the applaunchers fail to
resolve {target_engine} tokens in their icon paths. Thus, in order to
properly resolve icon paths, we make the shotgun_cache_actions command
use an entity context if an entity ID is specified (not done yet,
currently always uses a hardcoded context). This allows us to have our
detailed cache next to the simpler one, without both systems stepping on
each other's toes.