Fallback method for older shotgun_cache_actions
There is a hard check in shotgun_cache_actions for the args to be of
length 2. However, further changes in this branch will pass an entity ID
in order to create a context specific to an entity (and thus fetch more
information). We first try the new version of shotgun_cache_actions
(with an entity ID) and fallback to the old version (2 args) if it
needed.