shotgun_cache_actions use context if entity ID set
If an entity_id is provided, we set the context of the tk-shotgun engine
that gets started. This specifically allows the applauncher apps to
expand the icon paths that have a {target_engine} token (e.g. that's the
default value for a Maya launcher). If no entity_id is provided (e.g.
when the shotgun website updates the cache), an empty context is used.