Don't support specific context when caching
Also changed the logic that extracts the cache tokens to be more
concise.

This means that the cache can not be built using more precise
information (a specific context from an entity ID). This simplifies the
code quite a bit, but prevents from using mechanisms that need a context
(e.g. {target_engine} token for applaunchers). The reasoning behind it
is that supporting {target_engine} would require deeper changes to the
way environments are picked, so they will be handled separately.
Therefore, we remove the code that accepted an entity_id to be passed
around when generating the cache.

The advantage of this is that we can also drop the "_detailed" caches
and keep using the same ones that the website uses.