#33467 correctly detect Windows platform from tank cmd on Windows 10
In the Git shell on Windows 10, `umask` returns `MINGW64_NT-10.0` which we don't correctly detect as Windows in our `tank` shell scripts. This means that Toolkit won't be able to find the correct python interpreter to use and will generate an error.

This change adds logic to include the `MINGW64_NT*` family of values as belonging to the Windows platform so that the the correct interpreter will be found in `config/core/interpreter_windows.cfg`