#33456 Tank Command to get commands of entity
Adds the get_entity_commands tank command (API only) to get the list of
possible commands that can be executed on certain entities. This list is
fetched by invoking the ./tank executable of another pipeline
configuration. It fetches it from the other project's cache (and updates
it first if needed).
Extends the shotgun_*.txt caches to contain more information about the
commands: description and icon paths.

When building the cache, any app that relies on a context being set will
have problems resolving some of its information (e.g. {target_engine}
token in an applauncher icon path will not resolve because it will fail
to pick an environment). This is a limitation that we accept for the
moment.

Closes #216.