#34198 Improved error messaging for duplicate config paths
This improves the error reporting around the use case when there
are more than one pipeline configuration having the same path.
This leads to ambiguity and toolkit already reports an error message,
but the message wasn't entirely clear.

This improves the error message to include project id and pipeline
config id which will help aid support troubleshooting the issue.

Closes #221