#34420 Fixes a regression in the context validation.
This fixes a regression that was introduced in
0faf0bd7773cb4483666f3e30a51475bf4a83674. The regression forced the
context to evaluate the context.user property which in turn caused
the context to connect to Shotgun to run a lookup. This is not
supported in all states where Toolkit is executing so when the
actions where being generated by the Shotgun engine, errors would
appear due to the user evaluation happening. Evaluating user also has
performance implications. The new code does lazy evaluation and only
checks the context when a constraint has been defined in a manifest.

Closes #228