Refactors descriptors into a low-level and high-level layer.
This splits up the descriptor objects into an IODescriptor object
which controls the transfer and file management of an asset, and the
Descriptor object, which holds high level properties.

The high level descriptor object is a class hierarchy with classes
for core, config and bundles (apps, engines, frameworks). This allows
different descriptor types to implement different properties and
logic while all sharing the same underlying transport mechanism.