#33014 Adds on-the-fly context changing support to Toolkit engines and apps.
Previously, the context that a toolkit engine and its apps were running in was immutable. Any change in context required the engine and all apps to be torn down and restarted, which is not fast enough to be practical in some cases. With the introduction of Nuke Studio to the family of DCC applications that are supported by Shotgun Toolkit, it became necessary to support a faster approach to changing contexts. Engines that support on-the-fly context changes can now accept a new context and will handle reuse of already-running apps that themselves support context changes.

Closes #218.