First attempt at backward compatibility for legacy bundle cache path
* verified that an upgrade from v0.17.x to v0.18.x will cause errors
  saying that bundles do not exist on disk
* added a method to construct the legacy bundle cache path given
  information about a bundle and it's locator type
* each of the pre-existing locator types now has an additional fallback
  path which is the legacy bundle cache path

The thing to note about these changes is that they required the bundle type
to be forwarded to the IODescriptors for the pre-existing locator types
(git and manual. app_store already had the type and dev didn't need it). The
new core didn't rely on the bundle type to build paths because everything
goes directly underneath the "app_store" or "git" directories. Older cores
use the "apps/app_store", "apps/git", "engines/app_store" organization. So
this is something that needs to be looked at by Manne to see if it is a
reasonable tradeoff for backward compatibility support or if there is
another way to glean the bundle type when constructing the legacy path.