[minor] Non-immutable descriptors are now always dirty flagged.
When the configuration wrapper performs a comparison to check if a
managed config is up to date, it used to just compare the descriptor
urls as part of the check. For non-immutable descriptors such as dev
or path descriptors, this is not enough; even though the path to the
descriptor stays the same, the code  may have changed. So now the
update check always returns that non-immutable descriptors requires
an update at bootstrap time when running a unmanaged configuration.