[minor] Adds base resolver logic to bootstrap
This adds a new, more basic resolver to the bootstrap.

The DefaultShotgunResolver that was used previously contains what is
likely to be the eventual bootstrap resolve logic; the resolver looks
in Shotgun to find a pipeline configuration matching the given
namespace, project and pipeline config name ('branch'). It then looks
for paths and if those are not found, it looks for a config uri in a
field, tries to resolve the descriptor that this points at and use.
In case no pipeline config is found, it falls back on the base config.

This resolver requires some changes to Shotgun; the likely
introduction of a namespace field and supporting resolver logic
so that we can distinguish between namespaces as part of the
resolve (we may also do more complex operations in the future where
several configs can be merged at runtime etc). It also assumes a
config uri field.

For the time being, I have added a simpler BaseConfigurationResolver,
which fulfills the needs of the autonomous, plugin-based bootstrap
that is currently needed. This resolver does not allow for any types
of runtime overrides - it just returns the base configuration.

closes #241