[refactor] Refactored the unmanaged/managed configuration concept. Note that this terminology is only used internally in the code. The terms are now changed to Installed and Automatic.
An automatic config is handled by the toolkit manager and is
automatically installed and kept up to date. An automatic config
cannot and should not be edited directly in place, toolkit assumes
that it "owns" the config and may overwrite changes at any time.
Toolkit may also increase performance by caching out parts of the
configuration into more efficient data structures and such caches are
not automatically invalidated if a user makes a manual edit.

An installed configuration on the other hand is what we have today, a
configuration  that has been installed into a particular location on
disk and that is edited directly by a user. Cache scenarios are
dialled down to assume that files can change at any time and this
config will never be automatically overwritten.