[minor] Fixed bugs in descriptor caching
Improved the caching of descriptor objects - instead of having a
__new__ method to magically handle it behind the scenes, it is now
instead just done as part of the factory method. This makes the code
a bit easier to read and since we have a factory method and a
single point in the code where things are being instantiated, this
feels like the natural place to handle any smarts around creation of
new items.

This also resolves some bugs:

- __init__ is no longer executed all the time. With the previous
  implementation it was always being called, even for cached objects
  which was confusing and led to bugs.

- I think with the previous implementation, we were caching "latest"
  descriptors, which I think theoretically should be ok, but I can
  see this leading to potentially subtle future bugs. The factory
  based caching made it very straight forward to choose which
  descriptor types to cache, and latest and mutable descriptor
  types are now excluded from the cache.