for #28796: copy core api files during push_configuration if appropriate In a localized config, there are files related to the core API that need to exist for the config to function. These files were not being copied over to the target config which caused the target config to raise the error "The Toolkit configuration path points to an invalid configuration".
This change will copy the necessary files only when both the source and target pipeline configurations are localized.

When the source is localized and the target is not localized, we don't copy these files. We do issue a warning saying that the files will not be copied over and provide a list of the files in question.