#34913 Support for sparse configs
These changes eliminate the requirement that all bundle settings must be
explicitly specified in an environment. The new behavior uses the default_value
specified in the bundle's info.yml manifest as a fallback at runtime when not
specified in the environment file.

Any tank commands (update, install, etc) that modify environment files will no
longer populate new settings and set their values to the manifest's
corresponding default value. Settings with no default value will still be added
to the environment after prompting the user for a value. It should be pointed
out that no existing settings will be removed from environment files with the
changes.

Closes #247.