#34715: Adds config syntax to source hooks from current engine
This change adds a new config syntax for hooks to allow them to be sourced from
within the current engine. The new syntax looks like this:

hook_setting: {engine}/my_hook.py

At runtime, the {engine} token will be evaluated to the hooks directory within
the currently running engine. This change makes it possible to package
DCC-specific behavior with the engine rather than in each of the individual
multi-apps.

NOTE: This is branched from 34913_sparse_configs branch as it modifies the same
files. This should be merged after that work is complete.

Closes #255