for #35708 add syntax for optional environment includes
Adds support for optional include files. In the includes section of an environment YAML file, any file path that starts with a `?` will be flagged as optional and if the file does not exist on disk, no error will be raised. This enables you to set a base include file and optionally override it with with another file if it exists. Previously, any included file that didn't exist on disk would raise an exception.