[refactor] Latest is no longer an explicit descriptor keyword.
Instead, it is explictly specified in places where a latest behaviour
is relevant. This is to avoid confusion around latest. Previously,
latest could be specified anywhere, but it didn't make sense everywhere
because even if you could determine what the latest version was, you
sometimes didn't have access to it. In addition, following 'latest'
of an app in a config doesn't always make sense as every change to a
version may result on configuration changes. in order to make this
clear, latest is now separated out and used only in some places.